[GPU][DT] Expand inner dimensions during materialization. (#18294)
Previously we used "innerTiles[0]xinnerTiles[1]" layout for the
materialized shape; treat it like opaque shape, and it relies on later
lowering to resolve the opaque type. However, it is quite tricky to
represent the core contraction op with multi_mma op. This revisions
removes those dummy reshapes, and generates the target shape directly.

The revision also removes unnecessary methods from
GPUMaterializeEncoding.cpp, since the MaterializeEncodingInfo carries
the information in the patch.

Signed-off-by: hanhanW <hanhan0912@gmail.com>