[Codegen][Common] Reorder layout analysis to preserve anchored to_layout.
The motivation behind this PR was to solve the issue where despite us
emitting to_layout op at certain locations, we ended up generating
redundant to_simd earlier in the graph. This results us to not have
control on where the layout conversion actually happens.

One example of this case is when we emit a to_layout conversion for FP8
Attention. We emit a to_layout right before the 2nd contract and right
after the truncf to FP8. Since we do enforcement(backward propagation)
first and then propagate (forward propagation), we ended up generating
two layout conversions, one the original one we emitted, the second one
determined by layout analysis and placed in the middle of a elemwise op
that is part of the softmax.

This forces us to do shuffles/layout resolution in FP32 which hurts the
performance. Additionally, this is also not where we(compiler writers)
intended the layout conversion to happen.

To solve this issue, we re-ordered the layout analysis to do forward
propagation first and then enforcement(backward propagation). This gives
layout set by forward propagation precedence, which makes the to_layout
emitting more intuitive and behave more as expected.

Additionally, we made modification to getAgreedLayout to handle elemwise
propagation where some operands are not expected to have layout such as
arith.select who's condition operand probably don't have layouts.

Signed-off-by: Stanley Winata <stanley.winata@amd.com>