[GPU] Add support for materializing contraction ops. (#18196)
To serve the needs of data-tiled contraction op, we can't simply use
regular mma intrinsic because they implement more interface methods. The
multi_mma op with data-tiled mma intrinsic should be much more simpler
because we already relayout (i.e., pack + tile swizzle) the matrix.
Thus, the revision introduces `IREEGPU_DataTiledMMAAttr` which gives the
hints to multi_mma op.

The new attribute implements minimal interface methods. Only
`getABCElementTypes`, `getMNKShape` and `getSubgroupSize` is
implemented. They are all trivial methods which provides intrinsic
details like
[amd_matrix_instruction_calculator](https://github.com/ROCm/amd_matrix_instruction_calculator)
provided. The implementation detail of `getABCVectorTypes` is not clear
at this moment. We can potentially implement it based on the needs of
distributing data-tiled multi_mma op to lanes. The other missing
implementation is `populateOperandOffsetsSizesStrides`. Both methods
will be used in the later lowering, and we need to implement them to
connect e2e pieces.

Different from regular MMA attribute, this attributes has data-tiling
concept which can model unrolling and interleaving layouts. In the
concept of data-tiling, we always unroll the parallel dimensions (i.e.,
M, N dimensions) to be outermost, and interleave the unrolled K
dimension. I.e., the unrolled K dimension becomes the innermost
dimension. The constraint can be relaxed based on data-tiling needs. The
additional information can be added to `parameters`.

With the new DataTiledMMAAttr, the revision teaches the multi_mma op to
take unrolling factor into accounts in verifier.

The revision sorts out the ambiguities of materialized ops in
data-tiling and switch MMAAttr to DataTiledMMAAttr in materialization
patterns.

Note: this is not merging into main branch. This is mainly for
prototype, and we will contribute it to main later on.

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>