[Codegen][GPU] Enable forall hoisting with other ops in loop
The existing hoisting pattern was too restrictive in that it required
no other operations within the surrounding `scf.for` loop to be present.
This extends the pattern by allowing operations that are only used with
implicit capture from within the `scf.forall` op to be present. This
improves the robustness of the pattern against small variations that can
come out of tiling/loop normalization in earlier passes.