[Codegen] Add patterns for folding away no-op slices
Adds a pattern to fold away no-op slices to
OptimizeTensorInsertExtractSlices that calls an upstream utility that
uses the ValueBoundsInterface to determine whether the sizes of
a `tensor.extract_slice`/`tensor.insert_slice` are no-ops.

This folding is a pass option because some other pipelines are sensitive
to insert/extract_slice structure.