[Codegen][GPU] Handle dynamic and unaligned cases in DerivedThreadConfig (#18281)
This adds a default set of tile sizes for any dynamically shaped
copy/linalg ops that just uses the preferred vector size based on the
element type bitwidth of the linalg op. This same logic might also be
worth applying in aligned cases, but this patch is opting not to change
pre-existing behavior without proper benchmarking.

Additionally cleans up the tiling tests.