[LLVMGPU][ROCm] Disable kernarg preloading on non-CDNA3 targets
This is not supported on CDNA1/CDNA2 devices which, depending on the
firmware, may fail at runtime when executing code that uses kernel
argument preloading.

I considered adding this to the `TargetWgp` attribute but decided
against it because of how rocm-specific this optimization is.

Tested manually for a bunch of targets and looking at the ISA,
including: `gfx90a`, `gfx940`, `gfx942`, and `gfx1100`.