Rework special model testing to avoid shared cache interference. (#18344)
This commit reworks the special model testing for sdxl and sd3 so that
different instances generating compiler artifacts don't write to the
same shared cache. This is a major problem because the things we are
running for each test are no longer consistent. On top of that,
benchmarking was picking up from this same cache, so there were a bunch
of failures leading to chaos. Now, the output artifacts are written to a
place local to each instance running, so there should be no attempts to
write to the same cache in between jobs.

We still could use a better longterm solution to avoid this problem for
our input artifacts (mlir, weights), but for now we are good as they
rarely change and we can be strategic with when we update them.
Huggingface seems like a good option to move toward

Fixes https://github.com/iree-org/iree/issues/18336

---------

Signed-off-by: saienduri <saimanas.enduri@amd.com>