Distribute a launcher for iree-opt with the compiler wheel.
We've been exporting the driver symbol for iree-opt since forever but never provided an entry-point for it. Distributing it gives some basic IR manipulation capabilities and access to passes we think are important to distribute.

Note that we make no promises about which passes get included in a distributed iree-opt. Currently that is just about "all of them", which comes with a large binary/build time cost. In the future, we can strip this down to some minimal set. But for the moment, we are already paying for things how they are and there is no sense denying access.

Signed-off-by: Stella Laurenzo <stellaraccident@gmail.com>