Simplify and generalize GPU data-tiled encodings (#18434)
* Changes to `IREEGPU_DataTiledMMAAttr`:
  * Add `unroll_{m,n,k}` parameters.
  * Drop custom builder/parser/printer code.
* Implement `getABCVectorTypes()` by falling back on
`MMAAttr::getABCVectorTypes()` since that is unchanged.
* Changes to `MaterializeEncodingInfo`:
* Drop `intrinsicSize`, too specific to particular levels of swizzling.
Also drop `innerTileShapes` and `srcRank` as redundant with the general
expand_shape array that I wanted to introduce anyway (next point).
* Introduce a single optional encapsulated `swizzle` containing just
generic expandshape and permutation arrays, so they are agnostic and
dont need to be generalized again.
* Changes to `MaterializeEncoding` logic:
* Just create the `expand_shape` and `tranpose` dictated by the
`swizzle` structure. So this is completely general and doesn't introduce
unit dimensions to fit a rigid shape structure anymore.
* The removal of the unit dims accounts for the bulk of the lit test
differences, particularly in cases where the unit dims used to prevent
folding. Now the folding happens, so the tests observe a single folded
`pack` or `unpack` in some cases where the `expand_shape` and
`transpose` were folded.
* This relies on an upstream fix,
https://github.com/llvm/llvm-project/pull/107271. This PR currently
cherry-picks it, so you will need `git submodule update`.
  * Generalize some places that were hardcoding 2D matrices.
* Generally drop lots of logic that where inferring kernel shapes that
are now consistently something that the user passes, not something that
we compute (in the diff, search for `std::sqrt` to see the kind of code
being dropped here).
* Changes to `multi_mma` op:
  * Update the verifier.

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>