[Codegen][GPU] Add gpu configuration heuristics for attention (#18415)
MMA heuristics are refactored into utilities, and these utilities are
used to generate a schedule for attention. For now, attention still uses
the same heuristic as GEMM to calculate required tile sizes. This
heuristic will be improved in future (the heuristic for attention is to
have as many M tiles as possible).

This does not modify any tests, as MMA configuration logic is same.
Attention configuration logic is also same, but now has more checks for
shared memory and alignment for the QK matmul.