[Im2col] Add option to unroll decomposed im2col loops (#18342)
This adds an option to the DecomposeIm2colPass to unroll the resulting
loop nest of the decomposition, and sets it to true by default. This is
an easier form to handle in most lowerings of the op. In particular, it
is easier to generate better barrier placement on GPU lowerings, and
unrolling the loops results in better overall barrier placement with the
current pipeline.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>