[Codegen][LLVMGPU] Default to private memory space for scalar dispatches
Some scalar dispatches can fail to bufferize properly due to
extract_slice foldings. This should be fixed independently, but to at
least compile in such cases, it is fine to always default to private
memory allocations for the LLVMGPUTileAndFuse pipeline since all shared
memory allocations are required to be explicitly managed before
bufferization.