[Codegen][GPU] Add pass to resolve scf.forall ops (#18394)
This pass differs from the upstream pattern in that it supports dynamic
trip count forall loops as well as overprovisioning, however it has a
few restrictions with respect to indexing. It only supports linearized
thread indices, as well as loops where the iteration order matches the
order of fastest changing thread index (i.e. the outer most loop will be
assigned the slowest changing thread indices). These are practical
choices to avoid overcomplicating the implementation of this pass by
requiring us to normalize `scf.forall` operations before this pass. Such
a normalization pass is TODO.