[LLVMGPU][ROCm] Add validation on finalized llvm bitcode
Check that there are no unresolved external functions that will
otherwise compile fine but be rejected by the driver.

The validation happens on the llvm bitcode after bitcode linking, when
there is no chance for anything to resolve these external functions.

This is to guard against future issues similar to
https://github.com/iree-org/iree/issues/18534 that are close to
undebuggable for end users.