[GPU][DT] Add support for materializing tensor.empty and linalg.fill ops
The revisions moves the materialization patterns of tensor.empty and
linalg.fill to "populateShapeLikeMaterializeEncodingPatterns" set;
updates the comments. This set of patterns lower the ops with encodings
to the same op with materialized types.

It adds the tile swizzle shape inference to the tensor.empty pattern and
moves the utility to the "Utility methods" section without changes.

Signed-off-by: hanhanW <hanhan0912@gmail.com>