[Codegen][GPU] Force linalg ops to always use the same vector size
The current method of distributing linalg ops does not properly
guarantee coalesced memory accesses because the tile sizes beyond the
inner most are not distributed cyclically. This changes the tile sizes
for derived_thread_config to always pick the optimal vector size or
smaller for the inner most dims, and one for all outer dims. This will
turn into a loop that gets unrolled later on.