[LLVMGPU][ROCm] Move kernel annotation before serialization
ROCMTarget serialization is not the best location for this code because
it violates the following invariant in `buildLLVMGPUCodegenPassPipeline`:
>  - The module contains the final llvm.module ready to be serialized.

Specifically, requiring the ROCDL dialect to be loaded is problemantic
during serialization, as target-agnostic serialization in
(`iree-hal-serialize-executables`) does not register dependent dialects.

This PR moves kernel annotation just after conversion to ROCDL in the
`LowerToGPUPasses` pass pipeline.

Also add kernel annotation tests that were not straightforward to add
before, as annotation was not a freestanding pass whose output could be
inspected.

Signed-off-by: Jakub Kuderski <jakub@nod-labs.com>