[GPU][DT] Add support for materializing encoding ops with dynamic shape.
The revision uses OpFoldResult for shape calculation and tensor::EmptyOp
creation.

It also populates ResolveRankedShapedTypeResultDims patterns for folding
tensor.dim ops during materialization. It helps the cases that
tensor.dim ops are operated on unconverted ops, which makes logic
easier. E.g., it is used in unpack op creation, and the tensor.dim ops
are resolved to avoid illegal ops.

https://github.com/iree-org/iree/blob/e19950c4c01b20c329d22e729212d182da9fee76/compiler/src/iree/compiler/Codegen/Common/MaterializeEncodingIntoPackUnPack.cpp#L280-L281

Signed-off-by: hanhanW <hanhan0912@gmail.com>