[Codegen] Add pass for unrolling annotated for loops
This allows annotating for loops formed in earlier compilation stages to
be unrolled later on. The case this is used for today is for unrolling
loops from tiling producers of matmul operands (typically a copy).

If a loop has a dynamic trip count, the attribute will be dropped
silently (unrolling is best effort).