[DT] Collapse matmul_narrow_M/N field into round_dims_to attribute. (#18599)
The revision refines how we handle matmul_narrow_M/N in encodings. The
matmul_narrow_M/N attributes are retired in the patch. The information
is carried in `round_dims_to` attribute which is rounded to the next
power of two. E.g., if the value of `matmul_narrow_M` is `3`, it will be
`4` in the `round_dims_to` attribute.

The `kNarrowThreshold` constant is introduced to Encoding namespace. A
dimension is considered as "narrow" only if the size of the dimension is
less than `kNarrowThreshold`.

The revision implements `getMatmulNarrowM` and `getMatmulNarrowN`
helpers based on the definition of `round_dims_to` and the
`kNarrowThreshold` variable. Users can call the helpers to get the
narrow size of M/N. Again, you'll get the value that is rounded to the
next power of two if they are considered "narrow". E.g.,
`getMatmulNarrowM()` returns `4` if `M` equals to `3`.

The revision also fixes a bug in `chooseMatmulTile` on CPU side. The
hostDefinedUpperBound is not updated in the narrow-N cases.

Fixes https://github.com/iree-org/iree/issues/17729
Fixes https://github.com/iree-org/iree/issues/15466

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>