[LLVMGPU] Add KernelConfig for data tiled multi_mma ops (#18623)
This PR adds the KernelConfig logic needed for data tiled multi_mma ops.
The PR does 2 things:

1. Adds config logic for iree_gpu.multi_mma ops with a DataTiledMmaAttr
kind. The logic here is simple, setting workgroup and reduction tile
sizes to 1, and computing the workgroup_size based on the
DataTiledMmaAttr.
2. Modify the logic for `getThreadTileSizesFromLoopRanges` to allow for
inner dimensions that are smaller than the `vectorSize` (preferred load
size). This is needed for data tiled kernels, since the swizzled inputs
often have inner vector sizes that are expanded out into multiple
dimensions as a result of K interleaving. The config logic will now fill
as much of the `vectorSize` with the inner dimensions as possible.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>