[CPU] Collapsing methods that query a config is enabled.
It just looks if the UnitAttr is set in the dictionary or not.

The `translationInfo` check is moved above LLVMCPUPipelineOptions setup
because some configs are set to `translationInfo`. We can bail out
earlier when it does not exist.

It is a follow-up for https://github.com/iree-org/iree/commit/903ab0a7cefec46ddade4278e6d547dc9e4598a8

Signed-off-by: hanhanW <hanhan0912@gmail.com>