[CPU] Switching to linalg::LinalgOp for MaterializeContractionOp pattern
It moves the logics to MaterializeContractionOp pattern; the other one
(i.e., MaterializeDPSOperation<linalg::GenericOp>) becomes
ShapeIndependent pattern. The former one is a CPU specfic pattern, and
the latter one is general after the shapes are resolved. It is not moved
to ShapeIndependent category because the support of broadcast_map is not
implemented yet.

It should be an NFC, so there are no new tests.

Signed-off-by: hanhanW <hanhan0912@gmail.com>