Rework util.assume.* ops to util.assume.int and base on attributes.
This arguably should have been done from the get-go but the need wasn't apparent until implementing propagation across executable boundaries. The key additional requirement which arises there is in relation to capturing annotations from multiple specializations in a way that can be actioned. Having one op which carries all of the information and can represent both multiple constraints and multiple operands solves for all needed cases.

Also standardized on umin/umax verbiage to align with the upstream dataflow terminology, which will be hooked up in the following patch.

Signed-off-by: Stella Laurenzo <stellaraccident@gmail.com>