Honor narrow dims in GPUMaterializeEncoding (#18622)
This in particular implements honoring narrow M, N dimensions. But it's
also replacing hard-coded unrolling factors with a general heuristic
producing the same values. If nothing else, it documents the rationale
behind the current values and gives clues to generalize to other cases
and to try other variations for higher performance.

---------

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>