[Im2col] Allow multiple batch, M, and K dimensions on im2col result (#18593)
This PR adds support for multiple M and K dimensions in the result of
the im2col op. New metadata is added for correctly tracking the offsets
into the M and K dimensions along the multiple dimensions. New `m_strides`
and `k_strides` fields are added to the op, which represent a basis for
linearizing the `m_offset` and `k_offset` fields.

The motivation for doing this is that flattening the M dimension can
create an expand_shape op consumer of the resulting matmul. This can
cause issues with fusion and distribution, so it is useful to be able to
keep the multiple M dimensions intact. This PR does not change any
behavior of Conv2DToIm2col pass, which will be done in a later PR.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>