Adding `iree_vm_context_fork` to fork a context. Contexts will be cloned with the same set of modules and in the same frozen state. All module state within the context will be cloned via a module-defined function to allow for selective reinitialization and shallow/deep cloning choices. Module initializers are not re-run on the child context but at the end of the fork all modules will receive an `IREE_VM_SIGNAL_FORK` notification and can selectively reinitialize any resources they want. Once forked the parent and child context lifetimes are independent and different modules can be loaded in each.
emitc and python modules do not yet support forking and a fork request
will fail if one is loaded in the context. It'd be possible to fork a
context without one of those modules loaded and then load them into
the new forked contexts afterward.

The compiler currently does nothing special for reinitializing after a
fork. We should probably change mutable global tensors to be
reinitialized by outlining the initializers for them and calling the
same initialization function during `__init` and the fork signal handler.
For now every global will be forked by reference.

Fixes #18747.