Adding `IREE_STATUS_INCOMPATIBLE`. (#18752)
This allows error handling code to detect cases where the program is
incompatible with the hosting environment. One day when status payloads
are implemented we could attach exactly why in a programmatically
accessible way but for now a hosting application can dump the whole
supported environment/topology/etc.

Progress on #18749.