[ROCM][NFC] Add option to control SLP vectorization in llvm optimizations (#18865)
We keep SLP vectorization off because it can mask perf issues or create
regressions. However, on ROCM what we have noticed is that we are
hitting issues in several untested paths in the AMDGPU llvm backend
because we dont have SLP vectorization.
Here is an example of an issue that we wouldn't hit if SLP vectorization
was turned on
https://github.com/iree-org/iree/issues/18798

In this PR we are still keeping the exisiting behavior but provide a
flag to toggle it so that we can do the required benchmarking and
analysis.

Signed-off-by: Nirvedh <nirvedh@gmail.com>
Signed-off-by: Elias Joseph <eljoseph@amd.com>