Use FetchContent for both pybind11 and nanobind. (#18872)
This avoids various pinning problems and ensures that the version that
IREE specifies at the top of a build is used consistently throughout.
The nanobind incantation was taken from shortfin. The pybind11
incantation was adapted from [a
comment](https://github.com/pybind/pybind11/issues/2817#issuecomment-766392922)
and extended to use find_package integration.

The latter is sufficient for MLIR's pybind11 finding to delegate to the
one set at the top level. The MLIR code for finding pybind11 is ancient
and should be modernized to use FetchContent and find_package
integration with a pinned version. This would ensure consistent interop
with the rest of the ecosystem.

---------

Signed-off-by: Stella Laurenzo <stellaraccident@gmail.com>
Co-authored-by: Marius Brehler <marius.brehler@amd.com>
Signed-off-by: Elias Joseph <eljoseph@amd.com>