[Codegen] Add pass to decompose pack unpack ops at dispatch boundaries (#18852)
This PR adds a wrapper pass around DecomposePackUnPackOps, which adds a
control function for decomposing only packs and unpacks whose reshapes
can be folded with dispatch tensor loads/stores.

This PR also removes the public pass constructor with a control
function, opting to use wrapper passes in place of constructing passes
with arbitrary control functions. This is better for creating simple bug
repros, since the control function is part of the pass.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>
Signed-off-by: Elias Joseph <eljoseph@amd.com>