Revert "[DispatchCreation] Extend multi-use producer fusion" (#18917)
The reverted commit does not handle when the "consumer" uses a value
defined above. See https://github.com/iree-org/iree/issues/18879 for the
original issue. This is causing issue with ~15 onnx models.

I have a PR (https://github.com/iree-org/iree/pull/18855) to fix this by
including values used in an ops region in the backwards slice, but It is
waiting on upstream changes to `getBackwardSlice`. Currently, the PR is
using a wrapper around `getBackwardSlice` to acheive the same effect,
but this will be updated once the upstream change lands
(https://github.com/llvm/llvm-project/pull/113478)

Reverts iree-org/iree#18551

---------

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>
Signed-off-by: Elias Joseph <eljoseph@amd.com>