[ROCM] Turn on SLP vectorization (#18949)
After doing review on the benchmarks we see that this is not causing any
model regression and we also did a benchmarking suite of attention
kernels for which in the past we thought this could cause a regression.
However, we didnt find any significant perf change. See
[here](https://docs.google.com/spreadsheets/d/102hYwdOGehmi_HhnLxHevwAVzMraKgKEjdTdUlJ4_TU/edit?usp=sharing)
As this is needed for correctness issue found in
https://github.com/iree-org/iree/issues/18798 we are turning on the SLP
vectorizer.

Signed-off-by: Nirvedh <nirvedh@gmail.com>
Signed-off-by: Elias Joseph <eljoseph@amd.com>