[DT] Fold tensor.extract_slice away when turning encodings into NOP.
For static cases, it happens natrually because all the shapes are
statically-known. In the dynamic case, it is not easy to idenfity the
pattern because it could be a chain of the linalg/tensor operations. See
below snippet for examples.

Thus, the revision adds a pattern which turns tensor.extract_slice op
away when the source was from unset_encoding op.

```mlir
%41 = arith.muli %0, %c4 : index
%42 = tensor.empty(%41) : tensor<?x16384xf32>
%43 = linalg.fill ins(%cst : f32) outs(%42 : tensor<?x16384xf32>) -> tensor<?x16384xf32>
%44 = linalg.matmul_transpose_b ins(%lhs, %rhs : tensor<?x16384xf16>, tensor<16384x16384xf16>) outs(%43 : tensor<?x16384xf32>) -> tensor<?x16384xf32>
%extracted_slice_11 = tensor.extract_slice %44[0, 0] [%41, 16384] [1, 1] : tensor<?x16384xf32> to tensor<?x16384xf32>
```

Signed-off-by: hanhanW <hanhan0912@gmail.com>