[LLVMGPU] Create GPU pipeline option for IGEMM (#18981)
This PR moves the IGEMM transformation out of executable configuration,
and creates a new GPU pipeline option for using IGEMM for convolutions.
The PR reworks the way lowering configs are selected on convolutions,
but the resulting configurations after the IGEMM transformation are
unchanged. The PR makes the following changes:

- Creates a new field in GPUPipelineOptionsAttr called
use_igemm_convolution. This new option indicates to the TileAndFuse
pipeline to add the ConvolutionToIGEMM pass.
- Adds lowering config selection logic for convolutions going down the
IGEMM path. The old way to select lowering configs was to transform into
im2col + matmul, and then use the TileAndFuse matmul lowering config
logic to set a configuration on the matmul. Now, the config is set on
the convolution directly, and when the IGEMM transformation happens, the
config is moved to the matmul.
- Refactor some of the TileAndFuse matmul config logic to reuse in the
convolution config logic.
- Remove LLVMGPUConvolutionToIGEMMPass. This pass is no longer
necessary, since we do not need a config function or control function.
- Removes various IGEMM related tests from the TileAndFuse test files,
and creates new tests for the IGEMM path.
- Adds some util functions in LinalgExt to help infer the iteration
domain of and mapping of the resulting matmul after the Conv2DToIm2col
transformation. This is used when setting the lowering config on
convolutions for IGEMM.

This change is largely NFC-ish. The main functional change here is that
IGEMM will no longer be used when MFMA intrinsics cannot be targeted,
but the IGEMM path is not performant when that is the case anyway.
Otherwise, the changes from this PR simply change the ordering of how we
select lowering strategies for IGEMM convolutions (and the pipeline
options that are used for IGEMM). This will be useful for applying
tuning configurations for IGEMM later on, since the transformation no
longer runs during configuration.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>