Apply new versioning scheme
This implements the new versioning scheme proposed with #18938. For the
compiler, the runtime, the TF compiler tools and the TFLite compiler
tools, `version_info.json` files to allow to set the patch level
individually for each package. Furthermore, a `version_info.json` is
added at root level to have a common version for a (future) meta
package, now mainly used to determine the tag when publishing release
candidates. This file might be moved in future and additional,
per-package tags might be added.