[Encoding] Add extract_slice semantics to unset_encoding ops.
The commits updates the op to capture dynamic dimensions when the result
type has dynamic shapes. Because the op has the extract_slice semantics
support in the commit. This is the opposite behavior versus set_encoding
op (that has padding semantic). Thus, it updates the verifier. An
illegal test is moved to roundtrip test because it is now a valid op.
The revision adds a new illegal op to the invalid.test instead.

It updates all the lit tests, and it adds the dynamic sizes when we
create UnsetEncoding ops in SetEncodingPass. This enforces the captured
sizes to match the following extract_slice op.

It is a step towards https://github.com/iree-org/iree/issues/18956 and
it will make cancelling encodings much easier.

How we materialize encodings is not updated yet, so it should be an
NFC in terms of final generated codes on data-tiling path.

Signed-off-by: hanhanW <hanhan0912@gmail.com>