[LLVMGPUVectorDistribute] Fix batch dimensions extraction for attention-like ops
Currently, the batch dimensions are extracted as the union
of dimensions present across Q & K & V. This is not correct
if one of the dims in inputs (Q,K and V) could be seen as broadcasting.

Therefore, this commit changes this to be:

B = Union ( Q & K & O , K & V & O ) where if parallel dimensions
common between both matmuls will be treated as batching dimensions.

Signed-off-by: Manupa Karunaratne <manupa.karunaratne@amd.com>