[NFC] Refactor AttentionOpDetail struct
Currently, AttentionOpDetail struct is meant
to accept *all* the indexing maps of an attention-like
operator and use presumed indices to obtain the required
maps from that.

Going forward, this could be fragile if we want
to obtain oMap as well because oMap does not have
a static index unlike Q, K and V input.

Therefore, this commit refactors AttentionOpDetail to
expect callers to provide exactly the maps it requires.
Thus, rely on the callers to provide those maps.

Signed-off-by: Manupa Karunaratne <manupa.karunaratne@amd.com>