[python][tuner] Add bindings for MMAIntrinsic (#19095)
Expose MMA intrinsics to python. This includes both the `MMAIntrinsic`
attribute and the enum values. In addition, expose element/vector types
and shapes from the MMA interface (but not the interface itself).

This is so that the tuner can use the compiler as the SoT for all the
intrinsics and use them to parse target properties and generate
constraints.

The enum values are kept opaque so that we do not have to duplicate the
definitions in TD and C API headers.