Add punet benchmarking to the regression suite (#19088)
This commit adds support to benchmark punet fp16/fp8 performance TOM.
This concludes adding all the necessary testing for the SDXL model. It
also switches the compilation of punet to use the spec file as it is
necessary for tuning at the current state of the project. I've also
updated the artifacts in azure, and this time using the date as part of
the azure link so everyone knows the time the artifacts were generated.
Nithin is working on implementing the spec file optimizations as part of
the compiler itself, so we can remove the usage of such files in the
future. All tests have timeouts now too and updated existing ones
because the CLI flag seems to be per test timeouts (not the whole pytest
command itself). Now, we can avoid hangs such as
https://github.com/iree-org/iree/actions/runs/11748746984/job/32734141414

---------

Signed-off-by: saienduri <esaimana@amd.com>
Signed-off-by: saienduri <saimanas.enduri@amd.com>
Co-authored-by: saienduri <esaimana@amd.com>