[Util][NFC] OptimizeIntArithmetic: reduce calls to `eraseState` (#19130)
This pass is causing long compilation times for llama3 405b (even when
cherry-picking https://github.com/llvm/llvm-project/pull/115399). The
majority of the time is spent in this one pass. The compilation times
improve when calling `eraseState` only when ops are deleted. This is
similar to the upstream listeners in `UnsignedWhenEquivalent.cpp` and
`IntRangeOptimizations.cpp`. It appears this function loops over all
`LatticeAnchors` on each invocation to find the one to delete, causing
it to be slow. My (nonrigorous) experiment showed a decrease from 18 min
to 3 min compile time. My main concern here would be this affecting
correctness, as I don't know if this has unaccounted for side effects.

Signed-off-by: Ian Wood <ianwood2024@u.northwestern.edu>