More cleanup in `IREEGPUAttrs`. (#19161)
* Remove nearly 1,000 lines of code from `IREEGPUAttrs.cpp`, plus some
methods out of `IREEGPUAttrs.td`. The bulk of that is code that was only
used in a single place and didn't need to be in this central location,
gets moved to its single user. A few helper functions were completely
dead code, and a few additional simplifications.
* Write up-to-date comments for core data structures, particularly the
all-important `SingleSubgroupLayout`.
* Within `IREEGPUAttrs.cpp`, group at the top of the file the code that
encodes the raw information about MMA intrinsics, i.e. the code that
needs to be updated when a new intrinsic is supported.

---------

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>