[GPU] Use affine.delinearize_index for MMA tiles and vector distribution
This commit updates some by-hand delinearizations in MMA tile generation and vector distribution to use `affine.delinearize_index` instead.

The main tricky thing here is that a lot of that MMA code would use `(id / stride) % size`, whereas delinearize's outputs all have the form `(id % stride) / nextStride`. In all the cases at issue, we could use a utility to convert arrays of sizes and strides to a permutation on a delinearization basis.

In order to not break existing tests, the trivial-loop detector had to be manually instrumented to support `delinearize_index` (and I got `util.assume.int` while I was there). (I suspect there're a few other cases, and that, long-term, that detector should be using one of the bounds interfaces, but that's not this PR)# This is a combination of 7 commits.