[DispatchCreation] Add CSE before canonicalization of `flow.dispatch.workgroups`.
The real change in this PR is to add CSE before canonicalization so
that operands of `flow.dispatch.workgroups` can be better de-duped.

Since this needed a pipeline test, noticed the pipeline test existed
in Flow and called the global optimization, dispatch creation and flow
pipelines. This PR moves the test to dispatch creation and only tests
that pipeline.

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>