[Codegen] Add pass to verify workgroup distribution
While general verification is not possible, when using `scf.forall` for
workgroup distribution we have the opportunity for basic verification
that all writes are located within the distributed loop. In particular,
if we have any workgroup level loops, any write to global memory outside
is assumed to be illegal.

This happens after bufferization because it is impossible to do this
verification before determining the memory space of every tensor in the
dispatch.

This pass is relatively lightweight (two walks, both of which should be
short) and so is on by default for every CPU and GPU pipeline.