[CPU][DT] Define boundary encodings and execution encodings.
The boundary operations (i.e., bindings, flow.tensor.load/store, etc)
has EncodingSolver attached. It defines the layout for inputs/outputs.

The encodings on the operations (e.g., compute ops) captures all the
original encoding fields and we do know that they will be executed on
the execution device, so we should be able to insert some ops which
bring it from layout_a to layout_b.

It adds the indexing maps back, we can revisit it later.

Signed-off-by: hanhanW <hanhan0912@gmail.com>