[runtime][python] Fix device array deepcopy when not mappable
If the device array is not mappable to the host the deepcopy would fail
and result in an exception.

To mitigate this in case it is not mappable we copy.

Implement the __deepcopy__ to avoid double copy when not mappable.
If deepcopy is done through __reduce__ then once we would copy form
device to host and another time inside the deepcopy implementation.

Signed-off-by: Boian Petkantchin <boian.petkantchin@amd.com>