[Encoding] Introduce "layouts" field to EncodingAttr.
The revision introduces an optional "layouts" field to EncodingAttr. It
is an array of attributes that describes the potential layouts on the
device. It is an array because a device could have several executable
targets. Note that it can be any attribute with encoding attribute
interface implementation. The expectation of the field is to bridge the
logics between host codes and device codes. If an attribute does not
implement the interface, it could be discarded anytime.

It is a step towards https://github.com/iree-org/iree/issues/17924

Signed-off-by: hanhanW <hanhan0912@gmail.com>