[Encoding] Introduce EncodingLayoutAttrInterface.
The revisions introduces a new attribute interface in the Encoding
dialect. The interface is used to query layout information needed to
materialize encoding attributes.

Any backend can implement the interface to interpret an encoding layout
based on their needs.

The current expectation of the interface is to propagate layout
information from backends to the host compliation or other targets. The
expectation can be adjusted as long as we identify additional needs for
encodings/layouts.

It is a step towards https://github.com/iree-org/iree/issues/17924

Signed-off-by: hanhanW <hanhan0912@gmail.com>