Increase strictness of global isel use for ROCM
Increase strictness while this is primarily for development instead of
robust production cases (though it is still quite robust).

This will prevent fallback from global-isel with unsupported inputs
to prevent a situation of attributing a witnessed change to improved
global isel results that are actually caused by not using global isel
or to prevent a lack of change caused by not actually utilizing global
isel.

In this form, unsupported inputs will fail to compile entirely.

Signed-off-by: Tres Popp <tpopp@brium.ai>