[DT] Switch MaterializeEncodingIntoNop to use EncodingLayoutAttr.
The revision introduces LayoutAttrInterface attribute interface and
implements the codegen attribute that discards the encodings for
lowering. It is used in the MaterializeEncodingIntoNop pass that turns
encodings into nop.

The layout attributes for other backends are not implemented, so the
GPU/CPU materialization pass does not take it into account.

The revision also implements a isIdentityLayout method for the
MaterializeEncodingInfo struct and adds an unit test for it.

Signed-off-by: hanhanW <hanhan0912@gmail.com>