[mlperf][pkgci] Update punet-fp8 with reduction dim as last dim
We have changes in sharktank that converts reduction dim of the custom
attention to be the fastest dimension. This makes it more uniform with
the FP16 and canonical attention form and hopefully makes optimization
gets called more easily down the line with this.

Additionally, this is to prefetch S.T we do not break the coming
sharktank/mlperf bots and runs.

Signed-off-by: Stanley Winata <stanley.winata@amd.com>