[Codegen][LLVM] Annotate pointers an noundef + nonnull, consants as noundef (#19309)
We know, as a matter of the API, that pointers passed into GPU functions
and the HAL VM entry point won't be null and will have a defined value.

Annotating these facts may unblock LLVM optimizations. In particuar,
when future commits will annotate ranges on the push constants, those
annotations will need a `noundef` to be used. The existing `noalias` may
also sometimes be turned off by the possibility of undefined arguments.